refactor: move alert types from AlertStore to useAlerts for improved organization and clarity
